fix build for 1.0.3

+++ build.sh
@@ -1,5 +1,15 @@
 #!/bin/bash
 
+if [[ -z "$GRAFANA_ACCESS_POLICY_TOKEN" ]]; then
+    echo "Missing \$GRAFANA_ACCESS_POLICY_TOKEN, please export token"
+    exit 1
+fi
+
+if [[ -z "$ROOT_URL" ]]; then
+    echo "Missing \$ROOT_URL, please export what url to sign for"
+    exit 1
+fi
+
 set -xe
 
 rm -rf dist sdague-mezmo-datasource
